refactor: consolidate sync/async source duplication into shared modules
