+++ lib/gameinsights/async_/steamcharts.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 from typing import Any
 
 import aiohttp
@@ -6,8 +5,9 @@ from bs4 import BeautifulSoup
 from bs4.element import Tag
 
 from gameinsights.async_.base import AsyncBaseSource
+from gameinsights.sources._parsers import transform_steamcharts
+from gameinsights.sources._schemas import _STEAMCHARTS_LABELS
 from gameinsights.sources.base import SourceResult, SuccessResult
-from gameinsights.sources.steamcharts import _STEAMCHARTS_LABELS
 from gameinsights.utils.async_ratelimit import async_rate_limited
 
 
@@ -91,48 +91,8 @@ class AsyncSteamCharts(AsyncBaseSource):
 
         return SuccessResult(success=True, data=data_packed)
 
-    @staticmethod
-    def _safe_span_text(element: Tag | None) -> str | None:
-        if element is None:
-            return None
-        span = element.span
-        if span is None:
-            return None
-        return span.get_text()
-
     def _transform_data(self, data: dict[str, Any]) -> dict[str, Any]:
-        game_name_text = data["game_name"].get_text()
-        active_24h = self._safe_span_text(data["peak_data"][1])
-        peak_active = self._safe_span_text(data["peak_data"][2])
-
-        monthly_active_player = []
-        for row in data.get("player_data_rows", []):
-            cols = [col.get_text(strip=True) for col in row.find_all("td")]
-            if len(cols) != 5:
-                self.logger.log(
-                    f"Unexpected row structure: expected 5 cells, got {len(cols)}",
-                    level="warning",
-                    verbose=True,
-                )
-                continue
-            month, avg_players, gain, percentage_gain, peak_players = cols
-            monthly_active_player.append(
-                {
-                    "month": datetime.strptime(month, "%B %Y").strftime("%Y-%m"),
-                    "average_players": float(avg_players.replace(",", "")),
-                    "gain": float(gain.replace(",", "")) if gain not in ("-", "") else None,
-                    "percentage_gain": (
-                        float(percentage_gain.replace("%", "").replace(",", "").strip())
-                        if percentage_gain not in ("-", "")
-                        else 0
-                    ),
-                    "peak_players": float(peak_players.replace(",", "")),
-                }
-            )
-
-        return {
-            "name": game_name_text,
-            "active_player_24h": int(active_24h) if active_24h else None,
-            "peak_active_player_all_time": int(peak_active) if peak_active else None,
-            "monthly_active_player": monthly_active_player,
-        }
+        return transform_steamcharts(
+            data,
+            log_fn=lambda msg: self.logger.log(msg, level="warning", verbose=True),
+        )
