+++ lib/gameinsights/sources/howlongtobeat.py
@@ -15,54 +15,18 @@
 # ---------------------------
 
 import json
-import re
-from typing import Any, NamedTuple, cast
+from typing import Any, cast
 
 import requests
 
+from gameinsights.sources._parsers import (
+    extract_hltb_game_data,
+    generate_search_payload,
+    transform_howlongtobeat,
+)
+from gameinsights.sources._schemas import _HOWLONGTOBEAT_LABELS, _SearchAuth
 from gameinsights.sources.base import SYNTHETIC_ERROR_CODE, BaseSource, SourceResult, SuccessResult
 from gameinsights.utils.ratelimit import logged_rate_limited
-
-
-class _SearchAuth(NamedTuple):
-    """Authentication data extracted from HLTB init endpoint.
-
-    Known fields (token, hp_key, hp_val) are explicit for clarity.
-    The extras dict carries any future auth fields HLTB adds to the
-    init response, forwarding them automatically to search requests.
-    """
-
-    token: str
-    hp_key: str  # value of hpKey from init response (used as body field name)
-    hp_val: str  # value of hpVal from init response
-    user_agent: str  # UA used for the init request — must match in subsequent requests
-    extras: dict[str, str]  # any additional non-token string fields
-
-
-_HOWLONGTOBEAT_LABELS = (
-    "game_id",
-    "game_name",
-    "game_type",
-    "comp_main",
-    "comp_plus",
-    "comp_100",
-    "comp_all",
-    "comp_main_count",
-    "comp_plus_count",
-    "comp_100_count",
-    "comp_all_count",
-    "invested_co",
-    "invested_mp",
-    "invested_co_count",
-    "invested_mp_count",
-    "count_comp",
-    "count_speed_run",
-    "count_backlog",
-    "count_review",
-    "review_score",
-    "count_playing",
-    "count_retired",
-)
 
 
 class HowLongToBeat(BaseSource):
@@ -290,108 +254,15 @@ class HowLongToBeat(BaseSource):
         if response.status_code != 200:
             return None
 
-        match = re.search(
-            r'<script id="__NEXT_DATA__".*?>(.*?)</script>',
+        return extract_hltb_game_data(
             response.text,
-            re.DOTALL,
+            game_id,
+            log_fn=lambda msg: self.logger.log(msg, level="debug", verbose=True),
         )
-        if match:
-            try:
-                next_data = json.loads(match.group(1))
-                # Navigate the nested structure safely
-                game_data = (
-                    next_data.get("props", {}).get("pageProps", {}).get("game", {}).get("data", {})
-                )
-                game_list = game_data.get("game")
-                if isinstance(game_list, list) and len(game_list) > 0:
-                    return cast(dict[str, Any], game_list[0])
-            except (json.JSONDecodeError, KeyError, IndexError) as exc:
-                self.logger.log(
-                    f"HLTB __NEXT_DATA__ parse failed for game {game_id}: {exc}",
-                    level="debug",
-                    verbose=True,
-                )
-
-        return None
 
     @staticmethod
     def _generate_search_payload(game_name: str) -> dict[str, Any]:
-        """Generate the search payload.
-
-        Args:
-            game_name: The game name to search for.
-
-        Returns:
-            The payload dict for the search request.
-        """
-        return {
-            "searchType": "games",
-            "searchTerms": game_name.split(),
-            "searchPage": 1,
-            "size": 1,
-            "searchOptions": {
-                "games": {
-                    "userId": 0,
-                    "platform": "",
-                    "sortCategory": "popular",
-                    "rangeCategory": "main",
-                    "rangeTime": {"min": 0, "max": 0},
-                    "gameplay": {
-                        "perspective": "",
-                        "flow": "",
-                        "genre": "",
-                        "difficulty": "",
-                    },
-                    "rangeYear": {"max": "", "min": ""},
-                    "modifier": "",
-                },
-                "users": {"sortCategory": "postcount"},
-                "lists": {"sortCategory": "follows"},
-                "filter": "",
-                "sort": 0,
-                "randomizer": 0,
-            },
-            "useCache": True,
-        }
+        return generate_search_payload(game_name)
 
     def _transform_data(self, data: dict[str, Any]) -> dict[str, Any]:
-        """Transform raw data into the expected format.
-
-        The HLTB API returns time values in seconds. We convert them to minutes
-        for consistency with other sources. Also, we map the base time labels
-        to their '_avg' counterparts which represent the average completion time
-        as shown on the website.
-
-        Args:
-            data: The raw game data from HLTB.
-
-        Returns:
-            Dict with only the valid labels, with time values converted to minutes.
-        """
-        result: dict[str, Any] = {}
-        time_labels = {
-            "comp_main",
-            "comp_plus",
-            "comp_100",
-            "comp_all",
-            "invested_co",
-            "invested_mp",
-        }
-        for label in self._valid_labels:
-            raw_value: Any = None
-
-            # Map time labels to their '_avg' counterparts (average completion time)
-            if label in ('comp_main', 'comp_plus', 'comp_100', 'comp_all'):
-                avg_label = f"{label}_avg"
-                raw_value = data.get(avg_label)
-            elif label in ('invested_co', 'invested_mp'):
-                avg_label = f"{label}_avg"
-                raw_value = data.get(avg_label)
-            else:
-                raw_value = data.get(label, None)
-
-            if raw_value is not None and label in time_labels:
-                result[label] = cast(int, raw_value) // 60
-            else:
-                result[label] = raw_value
-        return result
+        return transform_howlongtobeat(data)
